chore: eval dir
